pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-search"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "transformers>=4.40",
    "Pillow>=10.0",
    "lancedb>=0.6",
    "pyarrow>=14.0",
    "lance>=0.9",
    "pyyaml>=6.0",
    "tqdm>=4.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
watch = ["watchdog>=3.0"]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
]

[project.scripts]
semantic-search = "semantic_search.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["semantic_search*"]
